Version 2.00!!!
I rebuilt the whole website from the ground up, the most notable changes are: dark mode, the new look, animations, and the use of cookies. I find it likely that soon in the near future I'll release a better version, but I'm too tired for all of that right now lmao.